added some unit tests (undo/redo not completely working)

+++ meteor/Models/TextState.cs
@@ -3,9 +3,9 @@ namespace meteor.Models;
 public class TextState
 {
     public string Text { get; }
-    public long CursorPosition { get; }
+    public int CursorPosition { get; }
 
-    public TextState(string text, long cursorPosition)
+    public TextState(string text, int cursorPosition)
     {
         Text = text;
         CursorPosition = cursorPosition;
